#1 Autodetection of components' ratio
- Add autodetections for components' ratio
- Refactor project structure
- Remove unused code
- Write proper docs in README.md
