+++ cmd/cli/main.go
@@ -1,0 +1,7 @@
+package main
+
+import "go-blurhash-cli/internal/cli"
+
+func main() {
+	cli.Execute()
+}
