--- cmd/root.go
@@ -1,20 +1,0 @@
-package cmd
-
-import (
-	"github.com/spf13/cobra"
-)
-
-var (
-	rootCmd = &cobra.Command{
-		Use: "blurhash-cli",
-	}
-)
-
-// Execute executes the root command.
-func Execute() error {
-	return rootCmd.Execute()
-}
-
-func init() {
-	cobra.OnInitialize()
-}
