--- main.go
@@ -1,7 +1,0 @@
-package main
-
-import "go-blurhash-cli/cmd"
-
-func main() {
-	cmd.Execute()
-}
