--- pkg/blurhash/illegal_argument_error.go
@@ -1,13 +1,0 @@
-package blurhash
-
-type IllegalArgumentError struct {
-	detailedMessage string
-}
-
-func (e *IllegalArgumentError) Error() string {
-	return e.detailedMessage
-}
-
-func NewIllegalArgumentError(message string) *IllegalArgumentError {
-	return &IllegalArgumentError{detailedMessage: message}
-}
